pyproject.toml
[project]
name = "voca-py"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.119.1,<0.120.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.2.1,<0.3.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "punq (>=0.7.0,<0.8.0)",
    "sqlalchemy[asyncio] (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "google-auth (>=2.41.1,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "typer (>=0.20.0,<0.21.0)",
    "httpx[http2] (>=0.28.1,<0.29.0)",
    "google-genai (>=1.46.0,<2.0.0)"
]

[tool.poetry]
packages = [{ include = "src", from = "." }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "ruff (>=0.14.2,<0.15.0)",
    "rich (>=14.2.0,<15.0.0)"
]

[tool.ruff]
line-length = 100
fix = true
unsafe-fixes = true  # <--- pozwala na automatyczne usuwanie elementów z importów
extend-select = ["F401", "F841"]  # F401 - nieużywany import, F841 - nieużywana zmienna